// src/shared/message-types.ts
/**
 * Message port name for communication between extension contexts
 */
export const MESSAGE_PORT_NAME = 'wordstream_message_port';

/**
 * Message types for communication between components
 */
export enum MessageType {
  // System messages
  SYSTEM_INITIALIZED = 'SYSTEM_INITIALIZED',
  SYSTEM_GET_STATUS = 'SYSTEM_GET_STATUS',
  SYSTEM_STATUS = 'SYSTEM_STATUS',
  SYSTEM_ERROR = 'SYSTEM_ERROR',
  BACKGROUND_READY = 'BACKGROUND_READY',
  CONTENT_SCRIPT_READY = 'CONTENT_SCRIPT_READY',
  POPUP_READY = 'POPUP_READY',
  
  // Auth messages
  AUTH_SIGN_IN = 'AUTH_SIGN_IN',
  AUTH_SIGN_IN_RESULT = 'AUTH_SIGN_IN_RESULT',
  AUTH_SIGN_OUT = 'AUTH_SIGN_OUT',
  AUTH_SIGN_OUT_RESULT = 'AUTH_SIGN_OUT_RESULT',
  AUTH_GET_USER = 'AUTH_GET_USER',
  AUTH_GET_USER_RESULT = 'AUTH_GET_USER_RESULT',
  AUTHENTICATE = 'AUTHENTICATE',
  SIGN_IN_RESULT = 'SIGN_IN_RESULT',
  SIGN_OUT = 'SIGN_OUT',
  SIGN_OUT_RESULT = 'SIGN_OUT_RESULT',
  GET_AUTH_STATE = 'GET_AUTH_STATE',
  AUTH_STATE = 'AUTH_STATE',
  AUTH_STATE_CHANGED = 'AUTH_STATE_CHANGED',
  
  // Translation messages
  TRANSLATE_TEXT = 'TRANSLATE_TEXT',
  TRANSLATE_WORD = 'TRANSLATE_WORD',
  TRANSLATE_WORD_RESULT = 'TRANSLATE_WORD_RESULT',
  TRANSLATION_RESULT = 'TRANSLATION_RESULT',
  SAVE_WORD = 'SAVE_WORD',
  SAVE_WORD_RESULT = 'SAVE_WORD_RESULT',
  GET_WORDS = 'GET_WORDS',
  GET_WORDS_RESULT = 'GET_WORDS_RESULT',
  DELETE_WORD = 'DELETE_WORD',
  DELETE_WORD_RESULT = 'DELETE_WORD_RESULT',
  WORD_UPDATED = 'WORD_UPDATED',
  WORD_DELETED = 'WORD_DELETED',
  
  // Settings messages
  SETTINGS_GET = 'SETTINGS_GET',
  SETTINGS_GET_RESULT = 'SETTINGS_GET_RESULT',
  SETTINGS_SAVE = 'SETTINGS_SAVE',
  SETTINGS_SAVE_RESULT = 'SETTINGS_SAVE_RESULT',
  SAVE_SETTINGS = 'SAVE_SETTINGS',
  GET_SETTINGS = 'GET_SETTINGS',
  SETTINGS_UPDATED = 'SETTINGS_UPDATED',
  
  // Stats messages
  GET_STATS = 'GET_STATS',
  STATS_UPDATED = 'STATS_UPDATED',
  
  // Game stats
  GAME_STATS_GET = 'GAME_STATS_GET',
  GAME_STATS_GET_RESULT = 'GAME_STATS_GET_RESULT',
  GAME_STATS_UPDATE = 'GAME_STATS_UPDATE',
  GAME_STATS_UPDATE_RESULT = 'GAME_STATS_UPDATE_RESULT',
  GET_GAME_STATS = 'GET_GAME_STATS',
  SAVE_GAME_STATS = 'SAVE_GAME_STATS',
  
  // AI assistant messages
  GEMINI_QUERY = 'GEMINI_QUERY',
  GEMINI_QUERY_RESULT = 'GEMINI_QUERY_RESULT',
  GEMINI_RESPONSE = 'GEMINI_RESPONSE',
  ASSISTANT_STATE_UPDATED = 'ASSISTANT_STATE_UPDATED',
  GEMINI_GET_HISTORY = 'GEMINI_GET_HISTORY',
  GEMINI_SAVE_HISTORY = 'GEMINI_SAVE_HISTORY',
  GEMINI_CLEAR_HISTORY = 'GEMINI_CLEAR_HISTORY',
  GET_VIDEO_CONTEXT = 'GET_VIDEO_CONTEXT',
  
  // State synchronization
  STATE_UPDATED = 'STATE_UPDATED',
  
  // Service status
  GET_SERVICE_STATUS = 'GET_SERVICE_STATUS',
  SERVICE_STATUS = 'SERVICE_STATUS'
} 